requirements-lambda.txt
#pyproj==2.4.1
#shapely~=1.6.4.post2
boto3-utils==0.0.3
xmljson~=0.2.0
requests>=2.18.1
